learning_system_BE/requirements.txt
Flask==2.3.3
Flask-JWT-Extended==4.6.0
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.1
pyotp==2.8.0
marshmallow==3.21.1
email-validator==2.1.1
requests
flask-cors
flask-redis
werkzeug
psycopg2-binary==2.9.9 